pyproject.toml
[project]
dependencies = [
    "advanced-alchemy==1.0.1",
    "crawl4ai==0.5.0.post4",
    "google-generativeai==0.8.4",
    "litestar-saq==0.5.0",
    "litestar[standard]==2.15.1",
    "minio==7.2.15",
    "pre-commit==4.2.0",
    "psycopg==3.2.6",
    "psycopg-pool==3.2.6",
    "ultimate-sitemap-parser==1.3.0",
]
description = "Add your description here"
name = "temp"
readme = "README.md"
requires-python = ">=3.12"
version = "0.1.0"

[tool.ruff]
ignore = [
    "D103",
    "D100",
    "D101",
    "INP001",
    "TRY003",
    "EM101",
    "N812",
    "PGH003",
    "TC001",
    "EXE001",
    "CPY001",
    "EXE002",
    "RUF012",
    "D102",
    "D106",
    "PLR6301",
    "TC002",
]
line-length = 120
select = ["ALL"]
target-version = "py312"
